backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.1
sqlalchemy==2.0.36
pymysql==1.1.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
motor==3.5.1